Migrate display related methods to AIDL part 2
This migrates more display related methods with simple arguments from ISurfaceComposer.h to the new AIDL interface.

Bug: 219574942
Test: atest SurfaceFlinger_test libsurfaceflinger_unittest libgui_test

+++ libs/gui/include/private/gui/ComposerServiceAIDL.h
@@ -68,7 +68,8 @@ public:
         if (!displayId) return nullptr;
         sp<IBinder> display;
         binder::Status status =
-                mComposerService->getPhysicalDisplayToken(displayId->value, &display);
+                mComposerService->getPhysicalDisplayToken(static_cast<int64_t>(displayId->value),
+                                                          &display);
         return status.isOk() ? display : nullptr;
     }
 };
